+++ util/checkers/doc.go
@@ -1,0 +1,3 @@
+// Package checkers provides functions and structures to verify common types and aggregate
+// errors.
+package checkers